driver service

+++ happyride-driver-service/src/main/java/io/vividcode/happyride/driverservice/model/Vehicle.java
@@ -19,7 +19,7 @@ public class Vehicle extends BaseEntityWithGeneratedId {
 
   private String mode;
 
-  private Long year;
+  private Integer year;
 
   private String registration;
 }
